refactor: moved filterBreadcrumbs to its own file
