+++ tests/js/spec/views/replays/detail/console/utils.spec.jsx
@@ -1,4 +1,4 @@
-import {filterBreadcrumbs} from 'sentry/views/replays/detail/console';
+import {filterBreadcrumbs} from 'sentry/views/replays/detail/console/utils';
 
 const breadcrumbs = [
   {
